test(gcp-detector): properly isolate GcpDetectorIntegration.test.ts
